refactor: orgnize sourcecode in `src` directory
Since there is more than a file now by including the web platform code,
it is quite a mess to include them in config files like tsconfig, gitignore or npmignore.

I have managed to place the typescript code that should be compiled in a `src` directory so that we can have less configuration.